# swagger.yaml
openapi: 3.0.3
info:
  title: VBI-Learning API Documentation
  description: API documentation for VBI-Learning platform
  version: 1.0.0
servers:
  - url: https://vbi-learning-api.example.com
...
paths:
...
  /api/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password
      operationId: resetPassword
      requestBody:
        description: User's email, reset code, and new password
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: john@example.com
                code:
                  type: string
                  example: reset-token-123
                newPassword:
                  type: string
                  example: newsecret123
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset successful
        '400':
          description: Bad request or Invalid email or code
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid email or code
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: An error occurred while processing the request.
  /api/course/upload-image:
    post:
      tags:
        - Course
      summary: Upload image files
      operationId: uploadImageFiles
      requestBody:
        description: Image files to upload
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Image files uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully uploaded
                  response:
                    type: array
                    items:
                      type: object
                      properties:
                        key:
                          type: string
                          example: image-file-name.jpg
                        presignedUrl:
                          type: string
                          example: https://s3.example.com/image-file-name.jpg
        '400':
          description: Bad request or Missing files
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example:
                      Bad request: Missing files in the request
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal server error
  /api/course/create:
    post:
      tags:
        - Course
      summary: Create a new course
      operationId: createCourse
      requestBody:
        description: Course data to create
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Course Title
                description:
                  type: string
                  example: Course description Add other properties here
                image:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Course created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 615bd5475e63f4166cccd6d2
                  name:
                    type: string
                    example: Course Title
                  description:
                    type: string
                    example: Course description Other course properties
        '400':
          description: Bad request or Title is taken
          content:
            application/json:
              schema:
                type: string
                example: Title is taken
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Course create failed. Try again
...
  /api/course/publish/{courseId}:
    put:
      tags:
        - Course
      summary: Publish a course
      operationId: publishCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          description: ID of the course
      responses:
        '200':
          description: Course published successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 615c83df157e024c78e7a294
                  name:
                    type: string
                    example: Published Course Name
                  published:
                    type: boolean
                    example: true
        '400':
          description: Bad request or Publish course failed
          content:
            application/json:
              schema:
                type: string
                example: Publish course failed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
  /api/course/unpublish/{courseId}:
    put:
      tags:
        - Course
      summary: Unpublish a course
      operationId: unpublishCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          description: ID of the course
      responses:
        '200':
          description: Course unpublished successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 615c83df157e024c78e7a294
                  name:
                    type: string
                    example: Unpublished Course Name
                  published:
                    type: boolean
                    example: false
        '400':
          description: Bad request or Unpublish course failed
          content:
            application/json:
              schema:
                type: string
                example: Unpublish course failed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
  /api/check-enrollment/{courseId}:
    get:
      tags:
        - Course
      summary: Check if a user is enrolled in a course
      operationId: checkEnrollment
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          description: ID of the course
      responses:
        '200':
          description: Check enrollment status
          content:
            application/json:
              example:
                status: true
                course: { /* Course details */ }
        '401':
          description: Unauthorized
          content:
            application/json:
              example: Unauthorized
  /api/free-enrollment/{courseId}:
    post:
      tags:
        - Course
      summary: Enroll user in a course for free
      operationId: freeEnrollment
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          description: ID of the course
      requestBody:
        description: User data for enrollment
        required: true
        content:
          application/json:
            example:
              user: { /* User data */ }
      responses:
        '200':
          description: User enrolled successfully
          content:
            application/json:
              example:
                message: Congratulation! You have successfully enrolled!
                course: { /* Course details */ }
        '400':
          description: Bad request or Enrollment create failed
          content:
            application/json:
              example: Enrollment create failed
        '401':
          description: Unauthorized
          content:
            application/json:
              example: Unauthorized
  /api/user-courses:
    get:
      tags:
        - Course
      summary: Get courses enrolled by a user
      operationId: userCourses
      responses:
        '200':
          description: List of user's enrolled courses
          content:
            application/json:
              example:
                - { /* Course details */ }
                - { /* Course details */ }
        '401':
          description: Unauthorized
          content:
            application/json:
              example: Unauthorized
...
  /api/mark-incomplete:
    post:
      tags:
        - Course
      summary: Mark a lesson as incomplete
      operationId: markIncomplete
      requestBody:
        description: Lesson and course data to mark as incomplete
        required: true
        content:
          application/json:
            example:
              courseId: 615c83df157e024c78e7a294
              lessonId: 615c83df157e024c78e7a295
      responses:
        '200':
          description: Lesson marked as incomplete
          content:
            application/json:
              example:
                ok: true
        '400':
          description: Bad request or Mark Lesson Incomplete Failed
          content:
            application/json:
              example: Mark Lesson Incomplete Failed
...